% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_builder.R
\name{fury_exclude_missing}
\alias{fury_exclude_missing}
\title{Exclude Cases with Missing Values}
\usage{
fury_exclude_missing(builder, vars)
}
\arguments{
\item{builder}{A \code{fury_spec_builder} object}

\item{vars}{Character vector. Variable names that must be non-missing}
}
\value{
Modified \code{fury_spec_builder} object
}
\description{
Exclude Cases with Missing Values
}
