% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_builder.R
\name{fury_flag_attention}
\alias{fury_flag_attention}
\title{Add Attention Check (Flag by Default)}
\usage{
fury_flag_attention(builder, var, pass_values, description, action = "flag")
}
\arguments{
\item{builder}{A \code{fury_spec_builder} object}

\item{var}{Character. Variable name of attention check}

\item{pass_values}{Numeric or character vector. Values that indicate passing}

\item{description}{Character. Description of the attention check}

\item{action}{Character. "flag" (default) or "exclude"}
}
\value{
Modified \code{fury_spec_builder} object
}
\description{
Add Attention Check (Flag by Default)
}
