% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_builder.R
\name{fury_flag_missing}
\alias{fury_flag_missing}
\title{Flag Cases with Missing Values (Don't Exclude)}
\usage{
fury_flag_missing(builder, vars)
}
\arguments{
\item{builder}{A \code{fury_spec_builder} object}

\item{vars}{Character vector. Variable names that must be non-missing}
}
\value{
Modified \code{fury_spec_builder} object
}
\description{
Flag Cases with Missing Values (Don't Exclude)
}
