% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_builder.R
\name{fury_partition_pretest}
\alias{fury_partition_pretest}
\title{Declare Pretest Partition}
\usage{
fury_partition_pretest(builder, date_var, start, end)
}
\arguments{
\item{builder}{A \code{fury_spec_builder} object}

\item{date_var}{Character. Name of date/datetime variable}

\item{start}{Character. Start date (YYYY-MM-DD or YYYY-MM-DD HH:MM:SS)}

\item{end}{Character. End date (YYYY-MM-DD or YYYY-MM-DD HH:MM:SS)}
}
\value{
Modified \code{fury_spec_builder} object
}
\description{
Declare Pretest Partition
}
