% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_builder.R
\name{fury_source}
\alias{fury_source}
\title{Add Data Source to Spec}
\usage{
fury_source(builder, file, format = "spss")
}
\arguments{
\item{builder}{A \code{fury_spec_builder} object}

\item{file}{Character. Path to data file}

\item{format}{Character. Data format (default "spss")}
}
\value{
Modified \code{fury_spec_builder} object
}
\description{
Add Data Source to Spec
}
