% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_builder.R
\name{print.fury_spec_builder}
\alias{print.fury_spec_builder}
\title{Print Spec Builder Summary}
\usage{
\method{print}{fury_spec_builder}(x, ...)
}
\arguments{
\item{x}{A \code{fury_spec_builder} object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the builder object
}
\description{
Print Spec Builder Summary
}
